+++ spec/models/carto/user_creation_spec.rb
@@ -326,4 +326,16 @@ describe Carto::UserCreation do
       user_creation.next_creation_step until user_creation.finished?
     end
   end
+
+  describe '#initialize_user' do
+    it 'initializes users with http_authentication without organization' do
+      created_via = Carto::UserCreation::CREATED_VIA_HTTP_AUTENTICATION
+      user = FactoryGirl.build(:valid_user)
+      user.organization_id.should == nil
+      user_creation = Carto::UserCreation.new_user_signup(user, created_via)
+      initialized_user = user_creation.send(:initialize_user)
+      initialized_user.should_not be_nil
+      initialized_user.organization_id.should be_nil
+    end
+  end
 end