+++ spec/support/factories/users.rb
@@ -71,7 +71,7 @@ module CartoDB
 
     def create_user(attributes = {})
       user = new_user(attributes)
-      user.valid?.should == true
+      raise "User not valid: #{user.errors}" unless user.valid?
       # INFO: avoiding enable_remote_db_user
       Cartodb.config[:signups] = nil
       user.save