+++ NEWS.md
@@ -1,5 +1,6 @@
 3.12.2 (2016-05-15)
 -------------------
+* Support for HTTP Header authentication.
 * Update navigation in public pages.
 * New feature in organizations management: user quota slider.